Add instruction stake tokens

+++ programs/stake-vault/src/state/stake_config.rs
@@ -9,7 +9,7 @@ pub struct StakeConfig {
     pub reward_mint: Pubkey,
     pub base_yield_rate: u64, 
     #[max_len(5, 5)]
-    pub lock_durations: Vec<u32>, // Lock days
+    pub lock_durations: Vec<u64>, // Lock days
     pub total_staked: u64,
     pub total_positions: u64,
     pub bump: u8,
